+++ bin/tprism/loader.py
@@ -168,94 +168,30 @@ def load_explanation_graph(expl_filename: str, option_filename: Optional[str] =N
     return graph, tensor_shapes, flags
 
 
-class OperatorLoader:
-    """
-    This class is used to load custom operators
-    """
-    def __init__(self) -> None:
-        self.operators:Dict[str,Type[BaseOperator]] = {}
-        self.base_module_name = "tprism.op."
-        self.module = None
+class PluginLoader:
+    """Generic plugin loader.
 
-    # a snake case operator name to class name
+    Scans .py files in a directory and registers every subclass of
+    `base_class` found in them under its snake-case class name.
+    Subclasses set `base_class` and `base_module_name`.
+    """
+    base_class: type = object
+    base_module_name: str = ""
+
+    def __init__(self) -> None:
+        self.plugins: Dict[str, Any] = {}
+
+    # a snake case plugin name to class name
     def to_class_name(self, snake_str: str) -> str:
         components = snake_str.split("_")
         return "".join(x.title() for x in components)
 
-    # class name to a snake case operator name
+    # class name to a snake case plugin name
     def to_op_name(self, cls_name: str) -> str:
         _underscorer1 = re.compile(r"(.)([A-Z][a-z]+)")
         _underscorer2 = re.compile("([a-z0-9])([A-Z])")
         subbed = _underscorer1.sub(r"\1_\2", cls_name)
         return _underscorer2.sub(r"\1_\2", subbed).lower()
-
-    def get_operator(self, name: str) -> Type[BaseOperator]:
-        assert name in self.operators, "%s is not found" % (name)
-        cls = self.operators[name]
-        assert cls is not None, "%s is not found" % (name)
-        return cls
-
-    def load_all(self, path: str) -> None:
-        search_path = os.path.dirname(__file__) + "/" + path
-        for fpath in glob.glob(search_path + "*.py"):
-            self.load(fpath)
-
-    def load(self, fpath: str) -> None:
-        print("[LOAD]", fpath)
-        name = os.path.basename(os.path.splitext(fpath)[0])
-        module_name = self.base_module_name + name
-        module = importlib.machinery.SourceFileLoader(
-            module_name, fpath
-        ).load_module()
-        self.load_module(module)
-
-    def load_module(self, module):
-        for cls_name, cls in inspect.getmembers(module, inspect.isclass):
-            if issubclass(cls, tprism.op.base.BaseOperator):
-                print("[IMPORT]", cls_name)
-                op_name = self.to_op_name(cls_name)
-                self.operators[op_name] = cls
-        
-    def set_cls(self, op_name, cls):
-        self.operators[op_name] = cls
-        
-
-
-class LossLoader:
-    """
-    This class is used to load custom loss functions
-    """
-    def __init__(self) -> None:
-        self.module = None
-        self.base_module_name = "tprism.loss."
-        self.losses:Dict[str,Any] = {}
-
-    # a snake case operator name to class name
-    def to_class_name(self, snake_str):
-        components = snake_str.split("_")
-        return "".join(x.title() for x in components)
-
-    # class name to a snake case operator name
-    def to_op_name(self, cls_name: str) -> str:
-        _underscorer1 = re.compile(r"(.)([A-Z][a-z]+)")
-        _underscorer2 = re.compile("([a-z0-9])([A-Z])")
-        subbed = _underscorer1.sub(r"\1_\2", cls_name)
-        return _underscorer2.sub(r"\1_\2", subbed).lower()
-
-    def get_loss(self, name: str) -> Tuple[Optional[Type[BaseLoss]], List[str]]:
-        m=re.match(r"^(.*)\(([0-9e\-\.]*)\)$", name)
-        if m:
-            # TODO: parer.pyを使って複数オプションに対応する
-            loss_name=m.group(1)
-            loss_params=[m.group(2)]
-        else:
-            loss_name=name
-            loss_params=[]
-        if name in self.losses:
-            cls = self.losses[name]
-            return cls, loss_params
-        else:
-            return None, loss_params
 
     def load_all(self, path: str) -> None:
         search_path = os.path.dirname(__file__) + "/" + path
@@ -275,15 +211,60 @@ class LossLoader:
         ).load_module()
         self.load_module(module)
 
-    def load_module(self, module):
+    def load_module(self, module) -> None:
         for cls_name, cls in inspect.getmembers(module, inspect.isclass):
-            if issubclass(cls, tprism.loss.base.BaseLoss):
+            if issubclass(cls, self.base_class):
                 print("[IMPORT]", cls_name)
                 op_name = self.to_op_name(cls_name)
-                self.losses[op_name] = cls
+                self.plugins[op_name] = cls
 
-    def set_cls(self,op_name,cls):
-        self.losses[op_name] = cls
+    def set_cls(self, op_name, cls) -> None:
+        self.plugins[op_name] = cls
+
+
+class OperatorLoader(PluginLoader):
+    """
+    This class is used to load custom operators
+    """
+    base_class = BaseOperator
+    base_module_name = "tprism.op."
+
+    @property
+    def operators(self) -> Dict[str, Type[BaseOperator]]:
+        return self.plugins
+
+    def get_operator(self, name: str) -> Type[BaseOperator]:
+        assert name in self.plugins, "%s is not found" % (name)
+        cls = self.plugins[name]
+        assert cls is not None, "%s is not found" % (name)
+        return cls
+
+
+class LossLoader(PluginLoader):
+    """
+    This class is used to load custom loss functions
+    """
+    base_class = BaseLoss
+    base_module_name = "tprism.loss."
+
+    @property
+    def losses(self) -> Dict[str, Type[BaseLoss]]:
+        return self.plugins
+
+    def get_loss(self, name: str) -> Tuple[Optional[Type[BaseLoss]], List[str]]:
+        m=re.match(r"^(.*)\(([0-9e\-\.]*)\)$", name)
+        if m:
+            # TODO: parer.pyを使って複数オプションに対応する
+            loss_name=m.group(1)
+            loss_params=[m.group(2)]
+        else:
+            loss_name=name
+            loss_params=[]
+        if loss_name in self.plugins:
+            cls = self.plugins[loss_name]
+            return cls, loss_params
+        else:
+            return None, loss_params
 
 def main():
     loss_loader = LossLoader()
